chore(RHINENG-15653): Cleanup RulesTable
