+++ src/Utilities/ruleHelpers.js
@@ -32,39 +32,6 @@ export const systemsWithRuleObjectsFailed = (systems) =>
     profiles: system.testResultProfiles,
   }));
 
-export const toRulesArrayWithProfile = (profilesWithRules) =>
-  profilesWithRules.flatMap((profileWithRules) =>
-    profileWithRules.rules.map(({ identifier, references, ...rule }) => {
-      const valueDefinitions = profileWithRules.valueDefinitions?.filter(
-        ({ id }) => {
-          return rule.values?.includes(id);
-        }
-      );
-
-      const ruleValues = Object.fromEntries(
-        Object.entries(
-          profileWithRules.ruleValues?.[profileWithRules.id] ||
-            profileWithRules.ruleValues?.[profileWithRules.profile.id] ||
-            {}
-        ).filter(
-          ([valueId]) =>
-            valueDefinitions.map(({ refId }) => refId).includes(valueId) ||
-            valueDefinitions.map(({ id }) => id).includes(valueId)
-        )
-      );
-
-      return {
-        ...rule,
-        references: references ? references : [],
-        identifier: identifier && identifier.label ? identifier : null,
-        profile: profileWithRules.profile,
-        policyId: profileWithRules.id,
-        valueDefinitions,
-        ruleValues,
-      };
-    })
-  );
-
 export const complianceScoreData = (profiles) => {
   const scoreTotal = profiles.reduce((acc, profile) => acc + profile.score, 0);
   const rulesPassed = profilesRulesPassed(profiles).length;
